+++ deploy/docker/django/initialimport.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+cd /var/www/productdb/source
+python3 manage.py initialimport $@